# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Focus",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        },
        {
            "url": "https://dev.api.focus.grupomobius.com"
        },
        {
            "url": "https://api.focus.grupomobius.com"
        }
    ],
    "paths": {
...
        "/roles/update/{id}": {
            "put": {
                "tags": [
                    "roles"
                ],
                "summary": "Updated role",
                "operationId": "updateRole",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id that to be updated",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Updated role object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Role"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Role"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Role not found"
                    }
                }
            }
        },
...
        "/api/vehicle-models/getall": {
            "get": {
                "tags": [
                    "vehicle-models"
                ],
                "summary": "Get all type vehicle models",
                "operationId": "2026a185f982e99b13b7b847f21c006e",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VehicleModel"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "An error has occurred."
                    }
                }
            }
        },
        "/vehicle-models/{id}": {
            "put": {
                "tags": [
                    "vehicle-models"
                ],
                "summary": "Updated vehicle model",
                "operationId": "updateRole",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id that to be updated",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Updated vehicle model object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/VehicleModel"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VehicleModel"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Vehicle Model not found"
                    }
                }
            }
        },
        "/zones/{id}": {
            "put": {
                "tags": [
                    "zones"
                ],
                "summary": "Updated zones",
                "operationId": "updateZone",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id that to be updated",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Updated zone object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Zone"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Zone"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Zone not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Role": {
                "title": "Role model",
                "description": "Role model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "description": {
                        "title": "Description",
                        "description": "Description",
                        "type": "string"
                    },
                    "created_at": {
                        "title": "Created at",
                        "description": "When was created",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "title": "Updated at",
                        "description": "When was last updated",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "VehicleModel": {
                "title": "Vehicle Model model",
                "description": "Vehicle Model model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "brand_id": {
                        "title": "Brand ID",
                        "description": "Brand ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "title": "Name",
                        "description": "Name",
                        "type": "string"
                    },
                    "created_at": {
                        "title": "Created at",
                        "description": "When was created",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "title": "Updated at",
                        "description": "When was last updated",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "Zone": {
                "title": "Zone model",
                "description": "Zone model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "campa_id": {
                        "title": "Campa ID",
                        "description": "Campa ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "title": "Name",
                        "description": "Name",
                        "type": "string"
                    },
                    "created_at": {
                        "title": "Created at",
                        "description": "When was created",
                        "type": "string"
                    },
                    "updated_at": {
                        "title": "Updated at",
                        "description": "When was last updated",
                        "type": "string"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}